refactor: src 하위 web 폴더 생성
